Simple Shuffle Implementations 2
This version is a basic version of a game of blackjack. So far what is implemented is the logic to draw any of the 52 cards at random. The game starts with a black screen but If you click the spacebar it acts as a hit me button and allows you to draw a card at random.  There is logic that handles the cards in the deck so that you cannot get a duplicate card so once a card is taken out of the pile the amount of cards in the remaining pool gets smaller and smaller until there are no cards left. If there are no cards left the game just resets to having no cards. This version's purpose was to serve as a showcase of the card assets and to get a randomized deck of shuffle cards every time you launch the game.